created user file within models package. Refactored User class into UserModel
