+++ resources/user.py
@@ -1,45 +1,6 @@
 import sqlite3
 from flask_restful import Resource, reqparse
-
-
-class User:
-    def __init__(self, _id, username, password):
-        self.id = _id
-        self.username = username
-        self.password = password
-
-    @classmethod
-    def find_by_username(cls, username):
-        con = sqlite3.connect('data.db')
-        cur = con.cursor()
-
-        query = "SELECT * FROM users WHERE username = ?"
-        result = cur.execute(query, (username,))
-        row = result.fetchone()
-        if row:
-            user = cls(*row)
-        else:
-            user = None
-        
-        con.close()
-        return user
-    
-    @classmethod
-    def find_by_id(cls, _id):
-        con = sqlite3.connect('data.db')
-        cur = con.cursor()
-
-        query = "SELECT * FROM users WHERE id = ?"
-        result = cur.execute(query, (_id,))
-        row = result.fetchone()
-        if row:
-            user = cls(*row)
-        else:
-            user = None
-        
-        con.close()
-        return user
-
+from models.user import UserModel
 
 class UserRegister(Resource):
     parser = reqparse.RequestParser()
@@ -59,7 +20,7 @@ class UserRegister(Resource):
     def post(self):
         data = UserRegister.parser.parse_args()
         
-        if User.find_by_username(data['username']):
+        if UserModel.find_by_username(data['username']):
             return {"message": f"User with username {data['username']} already exists."}, 409
         else: 
             con = sqlite3.connect('data.db')
